requirement.txt
fastapi
openai-agents
uvicorn
pydantic
python-dotenv
openai